Extract NavController extension methods to separate file
